requirements.txt
feedparser>=6.0.10
requests>=2.31.0
pydub>=0.25.1
flask>=3.0.0
python-dotenv>=1.0.0
